+++ alphastats/dataset/harmonizer.py
@@ -3,8 +3,8 @@
 from typing import Dict, Optional
 
 import pandas as pd
+from dataset.keys import Cols
 
-from alphastats.keys import Cols
 from alphastats.loader.BaseLoader import BaseLoader
