+++ alphastats/dataset/statistics.py
@@ -3,14 +3,14 @@ from typing import Dict, Tuple, Union
 
 import pandas as pd
 import pingouin
+from dataset.keys import Cols
+from dataset.utils import ignore_warning
 
-from alphastats.keys import Cols
 from alphastats.statistics.anova import Anova
 from alphastats.statistics.differential_expression_analysis import (
     DifferentialExpressionAnalysis,
 )
 from alphastats.statistics.multicova_analysis import MultiCovaAnalysis
-from alphastats.utils import ignore_warning
 
 
 class Statistics:
