+++ alphastats/gui/utils/analysis.py
@@ -7,8 +7,8 @@ from typing import Any, Dict, Optional, Tuple, Union
 import pandas as pd
 import streamlit as st
 from dataset.dataset import DataSet
+from dataset.keys import Cols, ConstantsClass
 
-from alphastats.keys import Cols, ConstantsClass
 from alphastats.plots.plot_utils import PlotlyObject
 from alphastats.plots.volcano_plot import VolcanoPlot
