+++ alphastats/gui/utils/ui_helper.py
@@ -3,10 +3,10 @@ import uuid
 
 import pandas as pd
 import streamlit as st
+from dataset.keys import ConstantsClass
 
 from alphastats import __version__
 from alphastats.gui.utils.preprocessing_helper import PREPROCESSING_STEPS
-from alphastats.keys import ConstantsClass
 
 # TODO add logo above the options when issue is closed
 # https://github.com/streamlit/streamlit/issues/4984
