+++ alphastats/plots/clustermap.py
@@ -3,9 +3,9 @@ from typing import Dict
 
 import pandas as pd
 import seaborn as sns
+from dataset.keys import Cols
 from dataset.statistics import Statistics
 
-from alphastats.keys import Cols
 from alphastats.plots.plot_utils import PlotUtils
