+++ alphastats/plots/dimensionality_reduction.py
@@ -5,9 +5,9 @@ import plotly
 import plotly.express as px
 import plotly.graph_objects as go
 import sklearn
+from dataset.keys import Cols
 from dataset.preprocessing import Preprocess
 
-from alphastats.keys import Cols
 from alphastats.plots.plot_utils import PlotlyObject, PlotUtils
 
 # make own alphastats theme
